updated middleware
