+++ spikes/data/CSE.Web.Middleware/CSE.Web.Middleware.csproj
@@ -2,8 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
-    <RootNamespace>CSE.Middleware</RootNamespace>
-    <AssemblyName>CSE.Middleware</AssemblyName>
+    <RootNamespace>CSE.Web.Middleware</RootNamespace>
     <VersionPrefix>0.0.1</VersionPrefix>
     <VersionSuffix>$([System.DateTime]::UtcNow.ToString(`MMdd-HHmm`))</VersionSuffix>
     <Copyright>Copyright (c) Microsoft Corporation. All rights reserved.</Copyright>
