+++ spikes/data/CSE.Web.Middleware/LoggerExtension.cs
