+++ spikes/data/CSE.Web.Middleware/loggerOptions.cs
