+++ spikes/data/CSE.Web.Middleware/stylecop.json
